fix(agent-work-lab): preserve identity-neutral source admission
