+++ .buildchain/kfd/kfd-3/collaboration-interface.prebuild.json
@@ -8,13 +8,13 @@
     "cwd": ".",
     "sourceSha": "f051741aceddda44a19942023650ac18906bb963",
     "registryPath": ".buildchain/kfd/kfd-3/surfaces.json",
-    "registryDigest": "sha256:b2d62b68057c56ad6568fe4e25f1d0e4590eea9e1a29d831c313cb1a1593c7ca"
+    "registryDigest": "sha256:3239328bcefec4bfd102c7e834d1b234e9e485c23af309d9ab89384f11b7819f"
   },
   "sourceRegistry": {
     "id": "kungfu",
     "path": ".buildchain/kfd/kfd-3/surfaces.json",
-    "sha256": "6da8b5fa9e459ae28f760996bdf4aa2921b52ea1554c5bffd0246e54d4c3a674",
-    "digest": "sha256:b2d62b68057c56ad6568fe4e25f1d0e4590eea9e1a29d831c313cb1a1593c7ca"
+    "sha256": "b4b61d0a8aea3b3f1cc2c52efd46689f8fbc81778938ca8f5ebfcb57847d9ee2",
+    "digest": "sha256:3239328bcefec4bfd102c7e834d1b234e9e485c23af309d9ab89384f11b7819f"
   },
   "upstreamKfd": {
     "aggregatePath": "developer/sdk/kfd/upstream-aggregate.json",
@@ -34,7 +34,7 @@
   "expectedArtifactVerification": {
     "command": "node scripts/buildchain-kfd-evidence.mjs --artifact-witness --json"
   },
-  "collaborationInterfaceDigest": "sha256:3af0a64475b80a667422d4587b98a4b598a1f59ae60d40026302383499b0340d",
+  "collaborationInterfaceDigest": "sha256:0813b77b27c376f146f7defdca7974854b61357f63281401b5aa5399d2523465",
   "collaborationInterface": {
     "schemaVersion": 1,
     "contract": "kungfu-buildchain-kfd-3-surface-registry",
@@ -161,26 +161,6 @@
         }
       },
       {
-        "id": "kungfu.agent-work-lab.autoplay",
-        "name": "kungfu agent-work-lab autoplay",
-        "kind": "cli-api-tui",
-        "participantProfile": "agent-or-developer",
-        "state": "declared",
-        "availability": "shipped",
-        "visibility": "public",
-        "participantFacing": true,
-        "public": true,
-        "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-        "evidencePath": "framework/core/src/python/kungfu/agent/commands.json",
-        "artifactPath": "framework/core/src/python/kungfu/agent/commands.json",
-        "maturity": "experimental",
-        "declaration": {
-          "owner": "kungfu",
-          "source": "scripts/buildchain-kfd-evidence.mjs",
-          "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json"
-        }
-      },
-      {
         "id": "kungfu.agent-work-lab.catalog",
         "name": "kungfu agent-work-lab catalog --json",
         "kind": "cli-api-gui",
@@ -244,46 +224,6 @@
         "id": "kungfu.agent-work-lab.plan",
         "name": "kungfu agent-work-lab plan --json",
         "kind": "cli-api-gui",
-        "participantProfile": "agent-or-developer",
-        "state": "declared",
-        "availability": "shipped",
-        "visibility": "public",
-        "participantFacing": true,
-        "public": true,
-        "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-        "evidencePath": "framework/core/src/python/kungfu/agent/commands.json",
-        "artifactPath": "framework/core/src/python/kungfu/agent/commands.json",
-        "maturity": "experimental",
-        "declaration": {
-          "owner": "kungfu",
-          "source": "scripts/buildchain-kfd-evidence.mjs",
-          "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json"
-        }
-      },
-      {
-        "id": "kungfu.agent-work-lab.starter-create",
-        "name": "kungfu agent-work-lab starter-create --destination <path> --expected-plan-root <root> --actor <actor> --execute --json",
-        "kind": "cli-api-gui-tui",
-        "participantProfile": "agent-or-developer",
-        "state": "declared",
-        "availability": "shipped",
-        "visibility": "public",
-        "participantFacing": true,
-        "public": true,
-        "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-        "evidencePath": "framework/core/src/python/kungfu/agent/commands.json",
-        "artifactPath": "framework/core/src/python/kungfu/agent/commands.json",
-        "maturity": "experimental",
-        "declaration": {
-          "owner": "kungfu",
-          "source": "scripts/buildchain-kfd-evidence.mjs",
-          "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json"
-        }
-      },
-      {
-        "id": "kungfu.agent-work-lab.starter-plan",
-        "name": "kungfu agent-work-lab starter-plan --json",
-        "kind": "cli-api-gui-tui",
         "participantProfile": "agent-or-developer",
         "state": "declared",
         "availability": "shipped",
@@ -3468,8 +3408,8 @@
     ],
     "audit": {
       "status": "passed",
-      "declared": 169,
-      "artifactPublic": 169,
+      "declared": 166,
+      "artifactPublic": 166,
       "policy": {
         "sourceOfTruth": ".buildchain/kfd/kfd-3/surfaces.json",
         "detectedButUnregistered": "product-specific-audit",
@@ -3582,26 +3522,6 @@
       }
     },
     {
-      "id": "kungfu.agent-work-lab.autoplay",
-      "name": "kungfu agent-work-lab autoplay",
-      "kind": "cli-api-tui",
-      "participantProfile": "agent-or-developer",
-      "state": "declared",
-      "availability": "shipped",
-      "visibility": "public",
-      "participantFacing": true,
-      "public": true,
-      "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-      "evidencePath": "framework/core/src/python/kungfu/agent/commands.json",
-      "artifactPath": "framework/core/src/python/kungfu/agent/commands.json",
-      "maturity": "experimental",
-      "declaration": {
-        "owner": "kungfu",
-        "source": "scripts/buildchain-kfd-evidence.mjs",
-        "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json"
-      }
-    },
-    {
       "id": "kungfu.agent-work-lab.catalog",
       "name": "kungfu agent-work-lab catalog --json",
       "kind": "cli-api-gui",
@@ -3665,46 +3585,6 @@
       "id": "kungfu.agent-work-lab.plan",
       "name": "kungfu agent-work-lab plan --json",
       "kind": "cli-api-gui",
-      "participantProfile": "agent-or-developer",
-      "state": "declared",
-      "availability": "shipped",
-      "visibility": "public",
-      "participantFacing": true,
-      "public": true,
-      "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-      "evidencePath": "framework/core/src/python/kungfu/agent/commands.json",
-      "artifactPath": "framework/core/src/python/kungfu/agent/commands.json",
-      "maturity": "experimental",
-      "declaration": {
-        "owner": "kungfu",
-        "source": "scripts/buildchain-kfd-evidence.mjs",
-        "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json"
-      }
-    },
-    {
-      "id": "kungfu.agent-work-lab.starter-create",
-      "name": "kungfu agent-work-lab starter-create --destination <path> --expected-plan-root <root> --actor <actor> --execute --json",
-      "kind": "cli-api-gui-tui",
-      "participantProfile": "agent-or-developer",
-      "state": "declared",
-      "availability": "shipped",
-      "visibility": "public",
-      "participantFacing": true,
-      "public": true,
-      "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json",
-      "evidencePath": "framework/core/src/python/kungfu/agent/commands.json",
-      "artifactPath": "framework/core/src/python/kungfu/agent/commands.json",
-      "maturity": "experimental",
-      "declaration": {
-        "owner": "kungfu",
-        "source": "scripts/buildchain-kfd-evidence.mjs",
-        "sourcePath": "framework/core/src/python/kungfu/agent/kfd3_api.registry.json"
-      }
-    },
-    {
-      "id": "kungfu.agent-work-lab.starter-plan",
-      "name": "kungfu agent-work-lab starter-plan --json",
-      "kind": "cli-api-gui-tui",
       "participantProfile": "agent-or-developer",
       "state": "declared",
       "availability": "shipped",
