+++ .buildchain/kfd/support-matrix.json
@@ -24,7 +24,7 @@
   },
   "kfd3Enforcement": {
     "schema": "kungfu-kfd-3-release-gate-enforcement/v1",
-    "declaredSurfaceCount": 169,
+    "declaredSurfaceCount": 166,
     "enforcedSurfaceCount": 0,
     "gateBindings": [],
     "claimBoundary": "Declared means discoverable; enforced requires a retained passed hard release Gate. Merely detected or declared surfaces are not enforced."
@@ -152,11 +152,11 @@
         "evidenceRoots": [
           {
             "path": ".buildchain/kfd/kfd-3/surfaces.json",
-            "sha256": "sha256:6da8b5fa9e459ae28f760996bdf4aa2921b52ea1554c5bffd0246e54d4c3a674"
+            "sha256": "sha256:b4b61d0a8aea3b3f1cc2c52efd46689f8fbc81778938ca8f5ebfcb57847d9ee2"
           },
           {
             "path": ".buildchain/kfd/kfd-3/capability-query.json",
-            "sha256": "sha256:150b956b83f61d8c59009d60e8b4bdc455420095aae77e0b9d7850be3b8a303b"
+            "sha256": "sha256:25f5e25473bfd37457e2dab306d8b764dad180b59bbc0d4201ee4e921ad5debf"
           }
         ]
       },
@@ -176,7 +176,7 @@
       },
       "claimClass": "release-qualified-support",
       "knownLimitations": [
-        "Support covers the 169 product-declared surfaces accepted by the exact Buildchain registry audit."
+        "Support covers the 166 product-declared surfaces accepted by the exact Buildchain registry audit."
       ],
       "owner": "buildchain",
       "nextGate": "Publish the exact product-declared registry audit and artifact closure in the Alpha release passport."
