+++ framework/gui/src/renderer/src/agent-work-lab.tsx
@@ -21,6 +21,19 @@ import {
   agentWorkLabRecommendation,
 } from '../../../../../extensions/agent-work-lab/experience/src/index.js';
 
+export function shouldOpenAgentWorkLab(
+  startupSurface: string,
+  availableKfxCount: number,
+): boolean {
+  return startupSurface === 'agent-work-lab' || availableKfxCount === 0;
+}
+
+export function unavailableKfxMessage(discoveredKfxCount: number): string {
+  return discoveredKfxCount === 0
+    ? 'no extensions found on the extension path'
+    : `${discoveredKfxCount} extensions discovered, but none admitted for GUI execution`;
+}
+
 export type AgentWorkLabMode = AgentWorkLabCaseId;
 
 type VisualStatus =
