+++ scripts/kfd-support-matrix.test.mjs
@@ -152,7 +152,7 @@ test('Shifu separates the immediate human verdict from stable agent JSON', () =>
   });
   assert.equal(human.status, 0, human.stderr);
   assert.match(human.stdout, /BOUNDED SUPPORT/);
-  assert.match(human.stdout, /169 declared, 0 enforced/);
+  assert.match(human.stdout, /166 declared, 0 enforced/);
   assert.doesNotMatch(human.stdout, /^\s*\{/u);
 
   const agent = spawnSync(SHIFU, ['kfd', 'status', '--json'], {
@@ -169,7 +169,7 @@ test('Shifu separates the immediate human verdict from stable agent JSON', () =>
     'KFD-3',
     'KFD-7',
   ]);
-  assert.equal(report.kfd3.enforcement.declaredSurfaceCount, 169);
+  assert.equal(report.kfd3.enforcement.declaredSurfaceCount, 166);
   assert.equal(report.kfd3.enforcement.enforcedSurfaceCount, 0);
 });
